requirements.txt
SQLAlchemy==2.0.25
Telethon==1.33.1
python-dotenv==1.0.0
tortoise-orm==0.20.0
aerich==0.7.2
pytonconnect==0.3.0
tonsdk==1.0.13
pytoniq-core==0.1.26
captcha==0.5.0
apscheduler==3.10.4
pytz==2023.3.post1